+++ src/actions/querystringsearch/querystringsearch.test.js
@@ -37,4 +37,22 @@ describe('querystringsearch action', () => {
       expect(action.request.data).toEqual({ ...data, b_size: 25, b_start: 25 });
     });
   });
+  it('should create an action to get the querystring results - context aware', () => {
+    const data = {
+      query: [
+        {
+          i: 'portal_type',
+          o: 'plone.app.querystring.operation.selection.any',
+          v: ['Document'],
+        },
+      ],
+    };
+    const action = getQueryStringResults('/folder1/folder2/object1', data);
+
+    expect(action.type).toEqual(GET_QUERYSTRING_RESULTS);
+    expect(action.request.op).toEqual('post');
+    expect(action.request.path).toEqual(
+      '/folder1/folder2/object1/@querystring-search',
+    );
+  });
 });